fixed precision.

+++ src/slic3r/GUI/Plater.cpp
@@ -5900,7 +5900,9 @@ void Plater::export_stl(bool extended, bool selection_only, bool zaxe_file_temp_
                 mesh.merge(mesh_to_export(*o, -1));
             }
             if (zaxe_file_temp_export) // center to model not the plate if within Zaxe file.
-                mesh.translate(Vec3f{-p->bed.build_volume().bed_center()[0],-p->bed.build_volume().bed_center()[1],0});
+                mesh.translate(
+                    Vec3f{static_cast<float>(-p->bed.build_volume().bed_center()[0]),
+                          static_cast<float>(-p->bed.build_volume().bed_center()[1]), 0});
         }
     }
     else {
